- Enable interface Claims - Several fixes

+++ src/Blazor.Extensions.WebUSB/USBDeviceFilter.cs
@@ -2,8 +2,8 @@ namespace Blazor.Extensions.WebUSB
 {
     public class USBDeviceFilter 
     {
-        public short? VendorId { get; set; }
-        public short? ProductId { get; set; }
+        public int? VendorId { get; set; }
+        public int? ProductId { get; set; }
         public byte? ClassCode { get; set; }
         public byte? SubClassCode { get; set; }
         public byte? ProtocolCode { get; set; }
